Started Vehicles
Having trouble with it.  Player should enter trigger volume, press a key and enter the car (FPC is turned off).  When the player wants to get out they press the same key and the car controller is disabled, the FPC/TPC is turned back on and placed next to the car facing forward (ExitSpot GO).  More later but this is the foundation for vehicles.  We can easily make a 4x4 and a truck like in Blackout from this.  Helicopter and boats would be different but probably not much.